+++ src/tests.rs
@@ -4,6 +4,7 @@ mod debug;
 mod decode_options;
 mod integers;
 mod limits;
+mod non_deterministic;
 mod parse;
 mod rundgren;
 mod sequence;